% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataregacc-package.R
\docType{package}
\name{dataregacc-package}
\alias{dataregacc}
\alias{dataregacc-package}
\title{dataregacc: A Data Package of Eurostat Regional Accounts data}
\description{
This data package consolidates in a single data.frame the different datasets released by Eurostat adding some useful dimensions (Country, NUTS level, etc).
}
\author{
\strong{Maintainer}: Luis Biedma \email{luis.biedma@ec.europa.eu}

}
\keyword{internal}
